+++ src/index.ts
@@ -1,3 +1,5 @@
 export { read } from './reader/read';
+export { readNMRiumObject } from './reader/readNMRiumObject';
+export { processJcamp } from './reader/readJcamp';
 export * from './migration/MigrationManager';
 // export { writeNmredata } from './writer/writeNmredata';