+++ src/reader/NmriumLikeObject.ts
@@ -2,7 +2,7 @@
 import type { Spectrum1D } from '../types/Spectra/Spectrum1D';
 import type { Spectrum2D } from '../types/Spectra/Spectrum2D';
 
-export interface Output {
+export interface NmriumLikeObject {
   spectra: Array<Spectrum1D | Spectrum2D>;
   molecules: Array<Record<string, any>>;
 }