+++ src/reader/UsedColors.ts
@@ -1,6 +1,0 @@
-export interface UsedColors {
-  //eslint-disable-next-line @typescript-eslint/naming-convention
-  '1d': string[];
-  //eslint-disable-next-line @typescript-eslint/naming-convention
-  '2d': string[];
-}