+++ src/reader/readNMReData.ts
@@ -1,20 +1,18 @@
 import { fileCollectionFromZip, FileCollectionItem } from 'filelist-utils';
 import { nmrRecordToJSON, getSDF } from 'nmredata';
 
+import { NmriumLikeObject } from '../types/NmriumLikeObject';
 import { NmredataParsingOptions } from '../types/Options/NmredataParsingOptions';
-import { Output } from '../types/Output';
 import { Source } from '../types/Source';
 import { isSpectrum2D } from '../utilities/tools/isSpectrum2D';
 import { addRanges } from '../utilities/tools/nmredata/addRanges';
 import { addZones } from '../utilities/tools/nmredata/addZones';
 
-import { UsedColors } from './UsedColors';
 import { readBruker } from './readBruker';
 import { readJcamp, readJcampFromURL } from './readJcamp';
 
 export async function readNMReData(
   file: FileCollectionItem,
-  usedColors: UsedColors,
   options: NmredataParsingOptions = {},
 ) {
   const fileCollection = await fileCollectionFromZip(await file.arrayBuffer());
@@ -24,17 +22,13 @@ export async function readNMReData(
 
   let { spectra, molecules = [] } = jsonData;
 
-  let nmrium: Output = {
+  let nmrium: NmriumLikeObject = {
     spectra: [],
     molecules,
   };
 
   for (const data of spectra) {
-    let { spectra: internalSpectra } = await getSpectra(
-      data.source,
-      usedColors,
-      options,
-    );
+    let { spectra: internalSpectra } = await getSpectra(data.source, options);
     for (const spectrum of internalSpectra) {
       const { info } = spectrum;
 
@@ -53,14 +47,14 @@ export async function readNMReData(
 
 async function getSpectra(
   source: Source,
-  usedColors: UsedColors,
   options: NmredataParsingOptions = {},
 ) {
   const { file, jcampURL } = source;
   const { brukerParsingOptions, jcampParsingOptions } = options;
 
   if (jcampURL) {
-    return readJcampFromURL(jcampURL, usedColors, {
+    return readJcampFromURL(jcampURL, {
+      name: file?.fileCollection.files[0].name,
       ...{ xy: true, noContours: true },
       ...jcampParsingOptions,
     });
@@ -70,13 +64,14 @@ async function getSpectra(
 
   switch (file.type) {
     case 'jcamp':
-      return readJcamp(file.fileCollection.files[0], usedColors, {
+      return readJcamp(file.fileCollection.files[0], {
+        name: file?.fileCollection.files[0].name,
         ...{ xy: true, noContours: true },
         ...jcampParsingOptions,
       });
     case 'brukerFiles':
-      return readBruker(file.fileCollection, usedColors, {
-        ...{ xy: true, noContours: true, keepOriginal: true },
+      return readBruker(file.fileCollection, {
+        ...{ xy: true, noContours: true },
         ...brukerParsingOptions,
       });
     default: