+++ src/types/Data2D.ts
@@ -1,5 +1,7 @@
+import { DoubleArray } from 'cheminfo-types';
+
 export interface Data2D {
-  z: Array<Array<number>>;
+  z: Array<DoubleArray>;
   minX: number;
   minY: number;
   minZ: number;