+++ src/types/NmriumLikeObject.ts
@@ -2,7 +2,7 @@
 import { Spectrum1D } from './Spectra/Spectrum1D';
 import { Spectrum2D } from './Spectra/Spectrum2D';
 
-export interface Output {
+export interface NmriumLikeObject {
   spectra: Array<Spectrum1D | Spectrum2D>;
   molecules: Array<Record<string, string>>;
 }