+++ src/types/Options/JcampParsingOptions.ts
@@ -3,8 +3,7 @@ import { Source } from '../Source';
 export interface JcampParsingOptions {
   name?: string;
   source?: Source;
-  noContours?: boolean;
-  xy?: boolean;
+  noContour?: boolean;
   keepRecordsRegExp?: RegExp;
   profiling?: boolean;
 }