+++ src/types/Options/ParsingOptions.ts
@@ -2,7 +2,7 @@ import { BrukerParsingOptions } from './BrukerParsingOptions';
 import { JcampParsingOptions } from './JcampParsingOptions';
 import { NmredataParsingOptions } from './NmredataParsingOptions';
 
-export interface Options {
+export interface ParsingOptions {
   name?: string;
   jcampParsingOptions?: JcampParsingOptions;
   brukerParsingOptions?: BrukerParsingOptions;