+++ CHANGELOG.md
@@ -1,5 +1,107 @@
 # Changelog
 
+## [0.5.2](https://github.com/cheminfo/nmr-load-save/compare/v0.5.1...v0.5.2) (2022-11-18)
+
+
+### Bug Fixes
+
+* improve integrationsum in jcamp-dx with assignment ([#58](https://github.com/cheminfo/nmr-load-save/issues/58)) ([6836b75](https://github.com/cheminfo/nmr-load-save/commit/6836b758fa84ef9dc30f3b1992f84903a2d87cd2))
+
+## [0.5.1](https://github.com/cheminfo/nmr-load-save/compare/v0.5.0...v0.5.1) (2022-11-18)
+
+
+### Bug Fixes
+
+* re/im for 2D FID from jcamp ([#56](https://github.com/cheminfo/nmr-load-save/issues/56)) ([00ea823](https://github.com/cheminfo/nmr-load-save/commit/00ea82355eee70a2eba53570644c429649bc2932))
+
+## [0.5.0](https://github.com/cheminfo/nmr-load-save/compare/v0.4.9...v0.5.0) (2022-11-17)
+
+
+### ⚠ BREAKING CHANGES
+
+* quadrants in 2D n load assignment from JCAMP-DX
+
+### Features
+
+* quadrants in 2D n load assignment from JCAMP-DX ([19b1ceb](https://github.com/cheminfo/nmr-load-save/commit/19b1cebebfb866173743749e9a8174c6656140fd))
+
+
+### Bug Fixes
+
+* formatSpectrum2D ([9ec2cb9](https://github.com/cheminfo/nmr-load-save/commit/9ec2cb9499fc8632938185a49d26c0ee09b7fe75))
+
+## [0.4.9](https://github.com/cheminfo/nmr-load-save/compare/v0.4.8...v0.4.9) (2022-11-10)
+
+
+### Bug Fixes
+
+* keep out internal variables in formatSpectra ([#45](https://github.com/cheminfo/nmr-load-save/issues/45)) ([81ac1b4](https://github.com/cheminfo/nmr-load-save/commit/81ac1b4620766eab25e5b059dcb13ed47fadcecc))
+
+## [0.4.8](https://github.com/cheminfo/nmr-load-save/compare/v0.4.7...v0.4.8) (2022-11-04)
+
+
+### Bug Fixes
+
+* keep all data from a nmrium object ([#41](https://github.com/cheminfo/nmr-load-save/issues/41)) ([dcaf99f](https://github.com/cheminfo/nmr-load-save/commit/dcaf99fe745e1b31078a0d23c8710d1e40a7f6b6))
+
+## [0.4.7](https://github.com/cheminfo/nmr-load-save/compare/v0.4.6...v0.4.7) (2022-10-28)
+
+
+### Bug Fixes
+
+* pass options to readNmriumObject ([#36](https://github.com/cheminfo/nmr-load-save/issues/36)) ([3311113](https://github.com/cheminfo/nmr-load-save/commit/3311113a1c0c30b115d12ae7cad8e96a53d18a71)), closes [#35](https://github.com/cheminfo/nmr-load-save/issues/35)
+
+## [0.4.6](https://github.com/cheminfo/nmr-load-save/compare/v0.4.5...v0.4.6) (2022-10-27)
+
+
+### Bug Fixes
+
+* info merge n filter out startwith dot files ([03122f9](https://github.com/cheminfo/nmr-load-save/commit/03122f9841ee6479c9e7f588f9118b2f13f7ebf1))
+
+## [0.4.5](https://github.com/cheminfo/nmr-load-save/compare/v0.4.4...v0.4.5) (2022-10-25)
+
+
+### Bug Fixes
+
+* **readNmriumObject:** merge info ([31767a5](https://github.com/cheminfo/nmr-load-save/commit/31767a5bf4544819e03c24f6f3c099478ccf2ad9))
+
+## [0.4.4](https://github.com/cheminfo/nmr-load-save/compare/v0.4.3...v0.4.4) (2022-10-25)
+
+
+### Bug Fixes
+
+* don't overwrite name from object or use title ([e9c4522](https://github.com/cheminfo/nmr-load-save/commit/e9c45229afca4dc0e1f5e89e127db20216119337))
+* ensure differents id when use cache ([7a29ae0](https://github.com/cheminfo/nmr-load-save/commit/7a29ae05bfaa8c2523081c0958e22987d70e00b6))
+
+## [0.4.3](https://github.com/cheminfo/nmr-load-save/compare/v0.4.2...v0.4.3) (2022-10-25)
+
+
+### Bug Fixes
+
+* readNMRiumObject ([#29](https://github.com/cheminfo/nmr-load-save/issues/29)) ([db9242e](https://github.com/cheminfo/nmr-load-save/commit/db9242e3642f4d7c5c59a00f9ef53ac74eb6617d))
+
+## [0.4.2](https://github.com/cheminfo/nmr-load-save/compare/v0.4.1...v0.4.2) (2022-10-20)
+
+
+### Bug Fixes
+
+* readNMRiumObject ([22e544c](https://github.com/cheminfo/nmr-load-save/commit/22e544c578ff462a674243718c8dfcc0958b9024))
+* update dependencies ([76b40a0](https://github.com/cheminfo/nmr-load-save/commit/76b40a05934db3013654993062d90c33dc8a7f7f))
+
+### [0.4.1](https://www.github.com/cheminfo/nmr-load-save/compare/v0.4.0...v0.4.1) (2022-10-19)
+
+
+### Bug Fixes
+
+* don't handle colors ([7db7f05](https://www.github.com/cheminfo/nmr-load-save/commit/7db7f055c79b2e323754b656862de73b35238059))
+
+## [0.4.0](https://www.github.com/cheminfo/nmr-load-save/compare/v0.3.0...v0.4.0) (2022-10-17)
+
+
+### Features
+
+* refactor n expose functions ([bc1edc0](https://www.github.com/cheminfo/nmr-load-save/commit/bc1edc0f79141ed0a6aa380e18202194f2d0f744))
+
 ## [0.3.0](https://www.github.com/cheminfo/nmr-load-save/compare/v0.2.2...v0.3.0) (2022-10-04)